Guest Control/VBoxService: Also made attribute allocation of VBOXSERVICECTRLSESSIONSTARTUPINFO + VBOXSERVICECTRLPROCSTARTUPINFO dynamic. [build fix] bugref:9320
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@84154 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/common/VBoxService/VBoxServiceControlProcess.cpp
@@ -92,7 +92,7 @@ int VgsvcGstCtrlProcessStartupInfoInitEx(PVBOXSERVICECTRLPROCSTARTUPINFO pStartu
     { \
         pStartupInfo->psz##a_Str = RTStrAlloc(a_cb); \
         AssertPtrBreak(pStartupInfo->psz##a_Str); \
-        pStartupInfo->cb##a_Str  = a_cb; \
+        pStartupInfo->cb##a_Str  = (uint32_t)a_cb; \
     }
 
     do
@@ -193,7 +193,7 @@ static PVBOXSERVICECTRLPROCSTARTUPINFO vgsvcGstCtrlProcessStartupInfoDup(PVBOXSE
     { \
         pStartupInfoDup->psz##a_Str = (char *)RTStrDup(pStartupInfo->psz##a_Str); \
         AssertPtrBreak(pStartupInfoDup->psz##a_Str); \
-        pStartupInfoDup->cb##a_Str  = strlen(pStartupInfoDup->psz##a_Str) + 1 /* Include terminator */; \
+        pStartupInfoDup->cb##a_Str  = (uint32_t)strlen(pStartupInfoDup->psz##a_Str) + 1 /* Include terminator */; \
     }
 
 #define DUP_MEM(a_Str) \
@@ -201,7 +201,7 @@ static PVBOXSERVICECTRLPROCSTARTUPINFO vgsvcGstCtrlProcessStartupInfoDup(PVBOXSE
     { \
         pStartupInfoDup->psz##a_Str = (char *)RTMemDup(pStartupInfo->psz##a_Str, pStartupInfo->cb##a_Str); \
         AssertPtrBreak(pStartupInfoDup->psz##a_Str); \
-        pStartupInfoDup->cb##a_Str  = pStartupInfo->cb##a_Str; \
+        pStartupInfoDup->cb##a_Str  = (uint32_t)pStartupInfo->cb##a_Str; \
     }
 
             DUP_STR(Cmd);
